requirements.txt
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
curl-cffi>=0.7
pytest>=8.0
